Moved SortHistory into a shared utils file
