+++ src/utils/index.ts
@@ -4,3 +4,4 @@ export * from './stringUtils';
 export * from './formUtils';
 export * from './formModeUtils';
 export * from './profileUtils';
+export * from './historyUtils';
